requirements.txt
flask
langchain
openai
faiss-cpu
python-dotenv
requests